Extract Sapling params prompt widget
